Publish this API's own configuration, and configure the bot
Ships deploy/kgsm-api.leaf.json — 69 settings in 13 groups — so the Control Panel
can show what the API itself is configured with. It declares itself read-only:
applying a change means restarting this service, which would kill the request
asking for it. That is a new leaf-level readOnly in the descriptor format, a
property of what the leaf is rather than of how the host was provisioned, so it
is declared instead of inferred from a missing drop-in. GET reports it with the
reason and PUT is a 409.

Adds a float field type for a setting that genuinely carries a fraction — a
similarity threshold, a sampling temperature. Coercing one through the integer
path would silently destroy it. Values are written in the invariant format so a
comma decimal separator cannot corrupt one.

Makes the bot a configuration target: its drop-in is installed and the polkit
grant covers its unit. The leaf-to-unit map is now explicitly a superset of the
leaves this API connects at runtime — the bot is configurable but not something
the API talks to.

Two fixes found by pointing the feature at real leaves. A leaf that ships a
descriptor without being in this API's built-in list now gets a real health
verdict after a config change; the canary resolved the unit from the built-in
list alone, so any such leaf had a working change rolled back. And a settings
file with comments is read as its floor, because the configuration stack that
reads it accepts them — a leaf whose annotated file it parses fine had its whole
floor reported as unknown over punctuation.

+++ src/Api/Services/Leaves/LeafConfigManifest.cs
@@ -30,10 +30,10 @@ public sealed record LeafConfigFieldDef(
 
     /// <summary>Inclusive bounds for a numeric field, mirroring the leaf's own parser floor so a value it
     /// would silently discard is rejected here instead — before any restart.</summary>
-    public long? Min { get; init; }
+    public double? Min { get; init; }
 
     /// <inheritdoc cref="Min"/>
-    public long? Max { get; init; }
+    public double? Max { get; init; }
 
     /// <summary>Display suffix (<c>ms</c>, <c>days</c>, <c>MB</c>). Presentation only.</summary>
     public string? Unit { get; init; }
